+++ README.md
@@ -49,5 +49,7 @@ yarn generate:openapi https://darrenvechain.github.io/thor-docs/thor.yaml
 
 Tests are executed using a custom private thor blockchain. The genesis of this blockchain is defined in `genesis.json`. This file contains accounts that are created with defined VET/VTHO balance for testing purposes. These accounts are then modelled in `wallet.ts`. To add a new account firstly add it to `genesis.json` and then to `wallet.ts`. 
 
+### Types of tests
 
+- thorest-api --> Tests that validate the swagger contract definition of the Thorest api
