+++ test/thorest-api/transactions/get-transaction.test.ts
@@ -1,5 +1,5 @@
-import { sendVetTransaction } from "../../src/transactions";
-import { Node1Client } from "../../src/thor-client";
+import { sendVetTransaction } from "../../../src/transactions";
+import { Node1Client } from "../../../src/thor-client";
 import assert from "node:assert"
